\name{dynamicTreeCut-package}
\alias{dynamicTreeCut-package}
\alias{dynamicTreeCut}
\docType{package}
\title{
Methods for detection of clusters in hierarchical clustering
dendrograms.
}
\description{
Contains methods for detection of clusters in hierarchical
clustering dendrograms.
}
\details{
\tabular{ll}{
Package: \tab dynamicTreeCut\cr
Version: \tab 1.11-3\cr
Date: \tab 2007-03-12\cr
Depends: \tab R, stats\cr
ZipData: \tab no\cr
License: \tab GPL version 2 or newer\cr
URL: \tab 
http://www.genetics.ucla.edu/labs/horvath/CoexpressionNetwork/BranchCutting/\cr
Packaged: \tab Wed Mar 12 12:32:38 2008; plangfelder\cr
Built: \tab R 2.4.1; ; 2008-03-12 12:48:47; unix\cr
}

Index:
\preformatted{
cutreeDynamic           Adaptive branch pruning of hierarchical
                        clustering dendrograms.
cutreeDynamicTree       Dynamic dendrogram pruning based on dendrogram
                        only
cutreeHybrid            Hybrid adaptive tree cut for hierarchical
                        clustering dendrograms.
indentSpaces            Spaces for indented output.
printFlush              Print arguments and flush the console.
treecut-package         Methods for detection of clusters in
                        hierarchical clustering dendrograms.
}
}
\author{
Peter Langfelder <Peter.Langfelder@gmail.com> and Bin Zhang
<binzhang.ucla@gmail.com>, with contributions from Steve
Horvath <SHorvath@mednet.ucla.edu>

Maintainer: Peter Langfelder <Peter.Langfelder@gmail.com>
}
\keyword{ package }
