\name{dynamicTreeCut-package}
\alias{dynamicTreeCut-package}
\alias{dynamicTreeCut}
\docType{package}
\title{
Methods for detection of clusters in hierarchical clustering
dendrograms.
}
\description{
Contains methods for detection of clusters in hierarchical
clustering dendrograms.
}
\details{
\tabular{ll}{
Package: \tab dynamicTreeCut\cr
Version: \tab 1.60\cr
Date: \tab 2013-07-24\cr
Depends: \tab R, stats\cr
ZipData: \tab no\cr
License: \tab GPL version 2 or newer\cr
URL: \tab 
http://www.genetics.ucla.edu/labs/horvath/CoexpressionNetwork/BranchCutting/\cr
}

Index:
\preformatted{
cutreeDynamic           Adaptive branch pruning of hierarchical
                        clustering dendrograms.
cutreeDynamicTree       Dynamic dendrogram pruning based on dendrogram
                        only
cutreeHybrid            Hybrid adaptive tree cut for hierarchical
                        clustering dendrograms.
indentSpaces            Spaces for indented output.
merge2Clusters          Merge two clusters
printFlush              Print arguments and flush the console.
treecut-package         Methods for detection of clusters in
                        hierarchical clustering dendrograms.
}
}
\author{
Peter Langfelder <Peter.Langfelder@gmail.com> and Bin Zhang
<binzhang.ucla@gmail.com>, with contributions from Steve
Horvath <SHorvath@mednet.ucla.edu>

Maintainer: Peter Langfelder <Peter.Langfelder@gmail.com>
}
\keyword{ package }
