\name{printFlush}
\alias{printFlush}
\title{Print Arguments and Flush the Console}
\description{Passes all its arguments unchaged to the standard \code{print} function; after the
execution of print it flushes the console, if possible.
}
\usage{
printFlush(\dots)
}
\arguments{
  \item{\dots}{Arguments to be passed to the standard \code{\link[base]{print}} function.}
}
\details{
  Passes all its arguments unchaged to the standard \code{print} function; after the
execution of print it flushes the console, if possible.
}
\value{Returns the value of the \code{\link[base]{print}} function.}
\author{
  Peter Langfelder, \email{Peter.Langfelder@gmail.com}
}
\seealso{
  \code{\link[base]{print}}
}
\keyword{print}
